docs: update README and project structure for open source release
- Changed project description to reflect open-source nature
- Updated features overview and key features
- Adjusted project structure for new layers (presentation, domain, data)
- Added contributing guidelines
- Included license information
- Updated contact information
