--- lib/core/databases/database_helper.dart
@@ -1,128 +1,0 @@
-import 'dart:async';
-
-import 'package:path/path.dart';
-import 'package:sqflite/sqflite.dart';
-import 'package:task_manager_app/features/authentication/data/models/user_model.dart';
-
-class DatabaseHelper {
-  static final DatabaseHelper _instance = DatabaseHelper._internal();
-
-  factory DatabaseHelper() => _instance;
-
-  DatabaseHelper._internal();
-
-  static Database? _database;
-
-  Future<Database> get database async {
-    if (_database != null) return _database!;
-    _database = await _initDatabase();
-    return _database!;
-  }
-
-  Future<Database> _initDatabase() async {
-    String path = join(await getDatabasesPath(), 'task_manager.db');
-    return await openDatabase(
-      path,
-      version: 1,
-      onCreate: _onCreate,
-    );
-  }
-
-  Future<void> _onCreate(Database db, int version) async {
-    await db.execute('''
-      CREATE TABLE users(
-        id TEXT PRIMARY KEY,
-        email TEXT NOT NULL,
-        username TEXT NOT NULL,
-        image TEXT,
-      )
-    ''');
-    await db.execute('''
-      CREATE TABLE settings(
-        key TEXT PRIMARY KEY,
-        value TEXT NOT NULL
-      )
-    ''');
-  }
-
-  // CRUD operations for UserModel
-  Future<void> insertUser(UserModel user) async {
-    final db = await database;
-    await db.insert(
-      'users',
-      user.toMap(), // Assuming toMap() converts UserModel to Map<String, dynamic>
-      conflictAlgorithm: ConflictAlgorithm.replace,
-    );
-  }
-
-  Future<UserModel?> getUser() async {
-    final db = await database;
-    final List<Map<String, dynamic>> maps = await db.query('users');
-    if (maps.isNotEmpty) {
-      return UserModel.fromMap(maps
-          .first); // Assuming fromMap() constructs UserModel from Map<String, dynamic>
-    }
-    return null;
-  }
-
-  Future<void> updateUser(UserModel user) async {
-    final db = await database;
-    await db.update(
-      'users',
-      user.toMap(),
-      where: 'id = ?',
-      whereArgs: [user.uid],
-    );
-  }
-
-  Future<void> deleteUser(String id) async {
-    final db = await database;
-    await db.delete(
-      'users',
-      where: 'id = ?',
-      whereArgs: [id],
-    );
-  }
-
-  // CRUD operations for settings
-  Future<void> insertSetting(String key, String value) async {
-    final db = await database;
-    await db.insert(
-      'settings',
-      {'key': key, 'value': value},
-      conflictAlgorithm: ConflictAlgorithm.replace,
-    );
-  }
-
-  Future<String?> getSetting(String key) async {
-    final db = await database;
-    final List<Map<String, dynamic>> maps = await db.query(
-      'settings',
-      where: 'key = ?',
-      whereArgs: [key],
-    );
-    if (maps.isNotEmpty) {
-      return maps.first['value'];
-    }
-    return null;
-  }
-
-  Future<void> updateSetting(String key, String value) async {
-    final db = await database;
-    await db.update(
-      'settings',
-      {'key': key, 'value': value},
-      where: 'key = ?',
-      whereArgs: [key],
-    );
-  }
-
-  Future<void> deleteSetting(String key) async {
-    final db = await database;
-    await db.delete(
-      'settings',
-      where: 'key = ?',
-      whereArgs: [key],
-    );
-  }
-}
