+++ lib/features/authentication/data/repos/auth_repo.dart
@@ -1,6 +1,5 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:task_manager_app/core/databases/database_helper.dart';
 import 'package:task_manager_app/core/di/service_locator.dart';
 import 'package:task_manager_app/features/authentication/data/models/login_with_email_and_password_request_body.dart';
 import 'package:task_manager_app/features/authentication/data/models/register_with_email_and_password_request_body.dart';
@@ -64,10 +63,7 @@ class AuthRepo {
   // Sign out
   Future<void> logOut() async {
     try {
-      final db = getIt<DatabaseHelper>();
-      await db.deleteUser(
-        _auth.currentUser!.uid,
-      );
+     
       await _auth.signOut();
     } catch (e) {
       rethrow;
