+++ lib/features/authentication/presentation/manager/cubits/google_sign_in/google_sign_in_state.dart
