+++ lib/features/authentication/presentation/manager/cubits/login_with_email_and_password/login_with_email_and_password_cubit.dart
@@ -8,7 +8,7 @@ import 'package:task_manager_app/core/errors/platform_exceptions.dart';
 import 'package:task_manager_app/core/utils/logger.dart';
 import 'package:task_manager_app/features/authentication/data/models/login_with_email_and_password_request_body.dart';
 import 'package:task_manager_app/features/authentication/data/repos/auth_repo.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/login_with_email_and_password/login_with_email_and_password_state.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/login_with_email_and_password/login_with_email_and_password_state.dart';
 
 class LoginWithEmailAndPasswordCubit
     extends Cubit<LoginWithEmailAndPasswordState> {
