+++ lib/features/authentication/presentation/manager/cubits/login_with_email_and_password/login_with_email_and_password_state.dart
