+++ lib/features/authentication/presentation/manager/cubits/registerwithemailandpassword/registerwithemailandpassword_state.dart
