+++ lib/features/authentication/presentation/views/login_view.dart
@@ -2,10 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:task_manager_app/core/constants/app_colors.dart';
-import 'package:task_manager_app/core/constants/app_constants.dart';
 import 'package:task_manager_app/core/constants/app_enums.dart';
 import 'package:task_manager_app/core/constants/app_images.dart';
-import 'package:task_manager_app/core/databases/database_helper.dart';
 import 'package:task_manager_app/core/di/service_locator.dart';
 import 'package:task_manager_app/core/helpers/spacing.dart';
 import 'package:task_manager_app/core/routes/routes.dart';
@@ -14,13 +12,14 @@ import 'package:task_manager_app/core/utils/utils.dart';
 import 'package:task_manager_app/core/widgets/app_button_widget.dart';
 import 'package:task_manager_app/core/widgets/app_divider_widget.dart';
 import 'package:task_manager_app/features/authentication/data/models/user_model.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/google_sign_in/google_sign_in_cubit.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/google_sign_in/google_sign_in_state.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/registerwithemailandpassword/register_with_email_and_password_cubit.dart';
-import 'package:task_manager_app/features/authentication/ui/widgets/register_form.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/google_sign_in/google_sign_in_cubit.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/google_sign_in/google_sign_in_state.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/login_with_email_and_password/login_with_email_and_password_cubit.dart';
+import 'package:task_manager_app/features/authentication/presentation/widgets/login_form.dart';
 
-class RegisterView extends StatelessWidget {
-  const RegisterView({super.key});
+class LoginView extends StatelessWidget {
+  const LoginView({super.key});
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -31,9 +30,8 @@ class RegisterView extends StatelessWidget {
             child: Column(
               children: [
                 BlocProvider(
-                  create: (context) =>
-                      getIt<RegisterWithEmailAndPasswordCubit>(),
-                  child: const RegisterForm(),
+                  create: (context) => getIt<LoginWithEmailAndPasswordCubit>(),
+                  child: const LoginForm(),
                 ),
                 verticalSpace(16),
                 const AppDividerWidget(),
@@ -43,25 +41,13 @@ class RegisterView extends StatelessWidget {
                   child: BlocConsumer<GoogleSignInCubit, GoogleSignInState>(
                     listener: (context, state) {
                       if (state is GoogleSignInFailure) {
-                        Utils.showSnackBar(
-                            context,
-                            'Register Failed: ${state.error}',
-                            SnackBarType.error);
+                        Utils.showSnackBar(context,
+                            'Login Failed: ${state.error}', SnackBarType.error);
                       } else if (state is GoogleSignInSuccess) {
                         Utils.showSnackBar(
-                            context,
-                            'Email Created Successfully!',
-                            SnackBarType.success);
-                        final db = getIt<DatabaseHelper>();
-                        db.insertSetting("isLoggedIn", "true");
-                        db.insertUser(UserModel.fromFirebase(state.user));
-                        Future.delayed(navigationDuration, () {
+                            context, 'Welcome Back!', SnackBarType.success);
                           Navigator.pushNamedAndRemoveUntil(
-                              context,
-                              Routes.homeView,
-                              arguments: state.user,
-                              (route) => false); // replace with your route
-                        });
+                            context, Routes.homeView, (route) => false);
                       }
                     },
                     builder: (context, state) {
@@ -71,7 +57,7 @@ class RegisterView extends StatelessWidget {
 
                       return AppButtonWidget(
                         hideIcon: true,
-                        text: 'Register with Google',
+                        text: 'Continue with Google',
                         onTap: () {
                           context.read<GoogleSignInCubit>().signInWithGoogle();
                         },
@@ -88,7 +74,7 @@ class RegisterView extends StatelessWidget {
                 verticalSpace(16),
                 AppButtonWidget(
                   hideIcon: true,
-                  text: 'Register with Apple',
+                  text: 'Continue with Apple',
                   onTap: () {
                     Utils.showSnackBar(
                         context, 'Coming Soon', SnackBarType.info);
@@ -104,15 +90,15 @@ class RegisterView extends StatelessWidget {
                 verticalSpace(32),
                 GestureDetector(
                   onTap: () {
-                    Navigator.pushNamed(context, Routes.loginView);
+                    Navigator.pushNamed(context, Routes.registerView);
                   },
                   child: Text.rich(
                     TextSpan(
-                      text: 'Already have an account? ',
+                      text: 'Don\'t have an account? ',
                       style: AppTextThemes.font12GreyRegular,
                       children: [
                         TextSpan(
-                          text: 'Login',
+                          text: 'Create Account',
                           style: AppTextThemes.font12WhiteRegular.copyWith(
                               color: AppColors.primaryColor.withOpacity(0.87)),
                         ),
