+++ lib/features/authentication/presentation/views/register_view.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:task_manager_app/core/constants/app_colors.dart';
+import 'package:task_manager_app/core/constants/app_constants.dart';
 import 'package:task_manager_app/core/constants/app_enums.dart';
 import 'package:task_manager_app/core/constants/app_images.dart';
-import 'package:task_manager_app/core/databases/database_helper.dart';
 import 'package:task_manager_app/core/di/service_locator.dart';
 import 'package:task_manager_app/core/helpers/spacing.dart';
 import 'package:task_manager_app/core/routes/routes.dart';
@@ -13,14 +13,13 @@ import 'package:task_manager_app/core/utils/utils.dart';
 import 'package:task_manager_app/core/widgets/app_button_widget.dart';
 import 'package:task_manager_app/core/widgets/app_divider_widget.dart';
 import 'package:task_manager_app/features/authentication/data/models/user_model.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/google_sign_in/google_sign_in_cubit.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/google_sign_in/google_sign_in_state.dart';
-import 'package:task_manager_app/features/authentication/logic/cubits/login_with_email_and_password/login_with_email_and_password_cubit.dart';
-import 'package:task_manager_app/features/authentication/ui/widgets/login_form.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/google_sign_in/google_sign_in_cubit.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/google_sign_in/google_sign_in_state.dart';
+import 'package:task_manager_app/features/authentication/presentation/manager/cubits/registerwithemailandpassword/register_with_email_and_password_cubit.dart';
+import 'package:task_manager_app/features/authentication/presentation/widgets/register_form.dart';
 
-class LoginView extends StatelessWidget {
-  const LoginView({super.key});
-
+class RegisterView extends StatelessWidget {
+  const RegisterView({super.key});
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -31,8 +30,9 @@ class LoginView extends StatelessWidget {
             child: Column(
               children: [
                 BlocProvider(
-                  create: (context) => getIt<LoginWithEmailAndPasswordCubit>(),
-                  child: const LoginForm(),
+                  create: (context) =>
+                      getIt<RegisterWithEmailAndPasswordCubit>(),
+                  child: const RegisterForm(),
                 ),
                 verticalSpace(16),
                 const AppDividerWidget(),
@@ -42,17 +42,22 @@ class LoginView extends StatelessWidget {
                   child: BlocConsumer<GoogleSignInCubit, GoogleSignInState>(
                     listener: (context, state) {
                       if (state is GoogleSignInFailure) {
-                        Utils.showSnackBar(context,
-                            'Login Failed: ${state.error}', SnackBarType.error);
+                        Utils.showSnackBar(
+                            context,
+                            'Register Failed: ${state.error}',
+                            SnackBarType.error);
                       } else if (state is GoogleSignInSuccess) {
                         Utils.showSnackBar(
-                            context, 'Welcome Back!', SnackBarType.success);
-                        final db = getIt<DatabaseHelper>();
-
-                        db.insertSetting("isLoggedIn", "true");
-                        db.insertUser(UserModel.fromFirebase(state.user));
-                        Navigator.pushNamedAndRemoveUntil(
-                            context, Routes.homeView, (route) => false);
+                            context,
+                            'Email Created Successfully!',
+                            SnackBarType.success);
+                        Future.delayed(navigationDuration, () {
+                          Navigator.pushNamedAndRemoveUntil(
+                              context,
+                              Routes.homeView,
+                              arguments: state.user,
+                              (route) => false); // replace with your route
+                        });
                       }
                     },
                     builder: (context, state) {
@@ -62,7 +67,7 @@ class LoginView extends StatelessWidget {
 
                       return AppButtonWidget(
                         hideIcon: true,
-                        text: 'Continue with Google',
+                        text: 'Register with Google',
                         onTap: () {
                           context.read<GoogleSignInCubit>().signInWithGoogle();
                         },
@@ -79,7 +84,7 @@ class LoginView extends StatelessWidget {
                 verticalSpace(16),
                 AppButtonWidget(
                   hideIcon: true,
-                  text: 'Continue with Apple',
+                  text: 'Register with Apple',
                   onTap: () {
                     Utils.showSnackBar(
                         context, 'Coming Soon', SnackBarType.info);
@@ -95,15 +100,15 @@ class LoginView extends StatelessWidget {
                 verticalSpace(32),
                 GestureDetector(
                   onTap: () {
-                    Navigator.pushNamed(context, Routes.registerView);
+                    Navigator.pushNamed(context, Routes.loginView);
                   },
                   child: Text.rich(
                     TextSpan(
-                      text: 'Don\'t have an account? ',
+                      text: 'Already have an account? ',
                       style: AppTextThemes.font12GreyRegular,
                       children: [
                         TextSpan(
-                          text: 'Create Account',
+                          text: 'Login',
                           style: AppTextThemes.font12WhiteRegular.copyWith(
                               color: AppColors.primaryColor.withOpacity(0.87)),
                         ),
