+++ lib/features/home/presentation/manager/logic/cubits/navigation_cubit/navigation_cubit_state.dart
