+++ lib/features/home/presentation/manager/logic/navigation/navigation_state.dart
