+++ lib/features/home/presentation/widgets/add_note_bottom_sheet.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:task_manager_app/features/home/ui/widgets/add_task_form.dart';
+import 'package:task_manager_app/features/home/presentation/widgets/add_task_form.dart';
 
 class AddTaskBottomSheet extends StatelessWidget {
   const AddTaskBottomSheet({super.key});
