+++ lib/features/onboarding/presentation/views/onboarding_view.dart
@@ -2,13 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:task_manager_app/core/constants/app_constants.dart';
 import 'package:task_manager_app/core/constants/app_strings.dart';
-import 'package:task_manager_app/core/databases/database_helper.dart';
-import 'package:task_manager_app/core/di/service_locator.dart';
 import 'package:task_manager_app/core/routes/routes.dart';
 import 'package:task_manager_app/core/widgets/app_button_widget.dart';
 import 'package:task_manager_app/core/widgets/text_button_widget.dart';
-import 'package:task_manager_app/features/onboarding/ui/widgets/on_boarding_dot_navigation.dart';
-import 'package:task_manager_app/features/onboarding/ui/widgets/onboarding_page_view.dart';
+import 'package:task_manager_app/features/onboarding/presentation/widgets/on_boarding_dot_navigation.dart';
+import 'package:task_manager_app/features/onboarding/presentation/widgets/onboarding_page_view.dart';
 
 class OnboardingView extends StatefulWidget {
   const OnboardingView({super.key});
@@ -91,15 +89,13 @@ class _OnboardingViewState extends State<OnboardingView> {
                       width: pageIndexNotifier.value == 2
                           ? 151.0.w
                           : double.infinity,
-                      onTap: () async{
+                      onTap: () async {
                         if (pageIndexNotifier.value != 2) {
                           controller.nextPage(
                             duration: duration,
                             curve: Curves.easeInOut,
                           );
                         } else {
-                          final db = getIt<DatabaseHelper>();
-                           await db.insertSetting('onboardingShown', 'true');
                           Navigator.pushReplacementNamed(
                             context,
                             Routes.registerView,
